Remove shop titles, move select titles out of select.
