+++ src/components/shop/shop.js
@@ -9,7 +9,7 @@ import { ScrollableList } from "./scrollable-list/scrollable-list.js";
 import RexUIPlugin from "../../../lib/rexuiplugin.min.js";
 import { getMetrics, onScaleChange } from "../../core/scaler.js";
 import { createConfirm } from "./confirm.js";
-import { createTitle } from "./shop-titles.js";
+import { createTitles } from "../../core/titles.js";
 import { createBalance } from "./balance-ui.js";
 import { createMenu } from "./menu.js";
 import { getSafeArea, getXPos, getYPos, getScaleFactor } from "./shop-layout.js";
@@ -38,7 +38,7 @@ export class Shop extends Screen {
             callback: this.back.bind(this),
         };
 
-        this.title = createTitle(this);
+        this.titles = createTitles(this);
         this.balance = createBalance(this);
 
         const inventoryFilter = item => item.id !== this.config.balance.value.key;
@@ -81,7 +81,6 @@ export class Shop extends Screen {
 
     setVisiblePane(pane) {
         Object.keys(this.panes).forEach(key => this.panes[key] && this.panes[key].setVisible(pane === key));
-        pane === "top" && this.title.setTitleText("Shop");
     }
 
     useCustomMessage() {
@@ -103,8 +102,6 @@ export class Shop extends Screen {
             this.panes.confirm = createConfirm(this, this.panes.confirm.title, this.panes.confirm.item);
         }
         const safeArea = getSafeArea(this.layout);
-        this.title.setScale(getScaleFactor({ metrics, container: this.title, fixedWidth: true, safeArea }));
-        this.title.setPosition(0, getYPos(metrics, safeArea));
         this.balance.setScale(getScaleFactor({ metrics, container: this.balance, safeArea }));
         this.balance.setPosition(
             getXPos(this.balance, safeArea, this.config.listPadding.x),
