+++ test/core/titles.test.js
@@ -3,8 +3,8 @@
  * @author BBC Children's D+E
  * @license Apache-2.0
  */
-import { createTitles } from "../../../src/components/select/titles.js";
-import * as text from "../../../src/core/layout/text-elem.js";
+import { createTitles } from "../../src/core/titles.js";
+import * as text from "../../src/core/layout/text-elem.js";
 
 describe("Select Screen - Titles", () => {
     let mockScene;
